Rewording of Step 3 in exercise 1 of Importing Data into OpenRefine episode
Someone I was working with found this step a bit confusing with and thought it would be better if they were directed to look at the box to make sure it was unchecked. I've found that during set up first time users are always a bit on edge about setting everything up right for the fear of doing it wrong and then having problems later in the lesson. Being as explicit as possible in setting things up can help alleviate any nerves they might have.