delete part of sentence that doesn't apply
The last clause in this sentence is missing a couple words, so it doesn't make sense. I also think this part of the sentence might be trying to say something about custom text facets, which are not referred to below. If this is the case, I am not sure this line about custom text facets is necessary.