Update 04-faceting-and-filtering.md
Added a column name for the 'Let's create a text facet' exercise. The wording here requires you to name a column. As a general comment, I noticed when prepping to teach this lesson that it tends to not mention a column when asking the instructor to demo a feature. I chose a column in advance to demo. Your approach (not naming a column) may be deliberate, so please feel free to reject this PR and rewrite in a non-column specific manner!