Add a dash to Formatted Date column
When we create this column, we should use dash, underscore, or CamelCase conventions when creating this column name to reflect what we teach learners about about in the Tidy Data lesson.  This is something I already do when teaching this lesson, but it should probably actually be in the lesson.